+++ src/ts/toolbar/Emoji.ts
@@ -1,5 +1,5 @@
 import emojiSVG from "../../assets/icons/emoji.svg";
-import {insertText} from "../editor/insertText";
+import {insertText} from "../sv/insertText";
 import {getEventName} from "../util/compatibility";
 import {getEditorRange, setSelectionFocus} from "../util/selection";
 import {insertHTML} from "../wysiwyg/insertHTML";