+++ src/ts/toolbar/MenuItem.ts
@@ -1,5 +1,5 @@
-import {insertText} from "../editor/insertText";
 import {i18n} from "../i18n/index";
+import {insertText} from "../sv/insertText";
 import {getEventName} from "../util/compatibility";
 import {updateHotkeyTip} from "../util/compatibility";
 import {toolbarEvent} from "../wysiwyg/toolbarEvent";