+++ CHANGELOG.md
@@ -58,7 +58,7 @@
 * [221](https://github.com/Vanessa219/vditor/issues/221) 输入复选框时出现乱码 `修复缺陷`
 * [222](https://github.com/Vanessa219/vditor/issues/222) The cursor does not enter when added in the middle of the list. `修复缺陷`
 * 文档更新
-  * 修改 `options.mode` 可选值为：'markdown', 'wysiwyg', 'typora'
+  * 修改 `options.mode` 可选值为：'sv', 'wysiwyg', 'ir'
   * toolbar 中的 wysiwyg 修改为 'edit-mode'
 
 ### v2.3.0 / 2020-03-12